[Refactor] Relocate to `domain.place.shared.dto`
