+++ src/main/java/boombimapi/domain/place/official/query/dto/request/ViewportRequest.java
@@ -1,10 +1,11 @@
 package boombimapi.domain.place.official.query.dto.request;
 
-import boombimapi.domain.place.official.query.dto.common.Coordinate;
+import boombimapi.domain.place.shared.dto.Coordinate;
 
 public record ViewportRequest(
     Coordinate topLeft,
-    Coordinate bottomRight
+    Coordinate bottomRight,
+    Coordinate memberCoordinate
 ) {
 
 }
